// admin-dashboard/src/app/core/api-client/model/shipmentResponse.ts
/**
 * CoreCourierService.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ParcelInfo } from './parcelInfo';
import { ContactInfo } from './contactInfo';


export interface ShipmentResponse { 
    id?: string | null;
    trackingNumber?: string | null;
    sender?: ContactInfo;
    receiver?: ContactInfo;
    parcel?: ParcelInfo;
    serviceType?: string | null;
    status?: string | null;
    estimatedDelivery?: string | null;
    createdAt?: string;
}