// admin-dashboard/src/app/core/api-client/model/shipmentResponsePagedResponse.ts
/**
 * CoreCourierService.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ShipmentResponse } from './shipmentResponse';
import { PaginationMeta } from './paginationMeta';


export interface ShipmentResponsePagedResponse { 
    data?: Array<ShipmentResponse> | null;
    pagination?: PaginationMeta;
}